+++ CHANGELOG.md
@@ -3,7 +3,8 @@
 ## Upcoming...
 
 - ... <!-- Add new lines here. Version number will be decided later -->
-- fix: Remove undefined `tempoarary` key from returned `KennitalaData` objects
+- feat: Add optional `separator` parameter to `formatKennitala`
+- fix: Remove undefined `temporary` key from returned `KennitalaData` objects
 
 ## 4.16.0